ai-agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.43.0
openai>=1.60.0
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.159.0
pypdf>=5.0.0
python-dotenv>=1.0.1
pydantic>=2.10.0
httpx>=0.28.0
